requirements.txt
tensorflow-gpu
numpy
scipy
nltk
pillow
distance
logging
multiprocessing
six
click
matplotlib
tensorboardX
imgaug
pathlib